% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveScore.R
\name{curveScore}
\alias{curveScore}
\title{Curve a Score}
\usage{
curveScore(itemScores, itemValues, itemWeights, specialScoring, lookup = NULL)
}
\arguments{
\item{itemScores}{a data frame with 1 row containing scores for a set of items}

\item{itemValues}{a data frame with 1 row containing max scores for a set of items}

\item{itemWeights}{a data frame with 1 row containing weights for a set of items}

\item{specialScoring}{a data frame with 1 row containing a scoring rule}
}
\value{
numeric of length 1 representing the curved score
}
\description{
Apply special scoring rule to a set of item scores to get an overall score
}
