% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadWorkbook2.R
\name{loadWorkbook2}
\alias{loadWorkbook2}
\title{Load Workbook 2}
\usage{
loadWorkbook2(file, xlsxFile = NULL, isUnzipped = FALSE)
}
\arguments{
\item{file}{filename of an xlsx file or path to an unzipped xlsx file}

\item{xlsxFile}{duplicate of \code{file}}

\item{isUnzipped}{logical, is \code{file} an xlsx file or a folder containing an unzipped xlsx file?}
}
\value{
Workbook object
}
\description{
Load an xlsx workbook object in original or unzipped form
}
