% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t.test2.R
\name{t.test2}
\alias{t.test2}
\title{t test 2}
\usage{
\method{t}{test2}(m1, m2, s1, s2, n1, n2, m0 = 0, equal.variance = FALSE,
  messageLevel = 0)
}
\arguments{
\item{m1}{mean of sample 1}

\item{m2}{mean of sample 2}

\item{s1}{the sample standard deviation of sample 1}

\item{s2}{the sample standard deviation of sample 2}

\item{n1}{the size of sample 1}

\item{n2}{the size of sample 2}

\item{m0}{the null value for the difference in means to be tested for.
Default is 0.}

\item{equal.variance}{whether or not to assume equal variance. Default is
FALSE.}

\item{messageLevel}{integer of length 1 indicating the level of diagnostic
messages to generate}
}
\value{
list of length 4 with elements "Difference of means", "Std Error",
  "t", and "p-value"
}
\description{
Perform a t test using summary stats rather than source data
}
\note{
This is based on code written by CrossValidated user Macro.  The
  original can be found at https://stats.stackexchange.com/a/30450
}
\examples{
x1 = rnorm(100)
x2 = rnorm(200)
# you'll find this output agrees with that of t.test when you input x1,x2
t.test2( mean(x1), mean(x2), sd(x1), sd(x2), 100, 200)
t.test(x1, x2)
}
