% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curveScore.R
\name{curveScore}
\alias{curveScore}
\title{Curve a Score}
\usage{
curveScore(itemScores, itemValues, itemWeights, specialScoring, lookup = NULL)
}
\arguments{
\item{itemScores}{a vector or a 1 row data frame containing whole number scores for a set of items}

\item{itemValues}{a vector or a 1 row data frame containing whole number max scores for a set of items}

\item{itemWeights}{a vector or a 1 row data frame containing whole number weights for a set of items}

\item{specialScoring}{a 1 row data frame containing a scoring rule}
}
\value{
numeric of length 1 representing the curved score (generally on a scale of 0 to 1)
}
\description{
Apply special scoring rule to a set of item scores to get an overall score
}
\examples{
# Sample Data 1
itemScores = c(1,2,1,1,2)
itemValues = c(2,2,2,2,3)
itemWeights = c(2,2,2,2,3)
specialScoring = data.frame(t(c("Extra Credit items", "9",NA,NA,NA,NA)), stringsAsFactors = F)
names(specialScoring) = c("function",paste0("parameter ",1:5))

# Test Run 1
curveScore(itemScores, itemValues, itemWeights, specialScoring)
}
