% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReport.R
\name{generateReport}
\alias{generateReport}
\title{Generate Report}
\usage{
generateReport(DataLocation = choose.dir(default = "J:/tests/2016-2017/"),
  ComparisonFileName = "comparison and topic alignment.xlsx",
  ReportFileName = "scores.xlsx", TMS = "LinkIt", SMS = "PowerSchool",
  UploadFilenames = c("upload_percentages.csv", "upload_totalpoints.csv"),
  template = NULL, messageLevel = 0)
}
\arguments{
\item{DataLocation}{a folder path}

\item{ComparisonFileName}{The name of the file with the test set up and comparison info}

\item{ReportFileName}{Desired name of the output excel file}

\item{TMS}{Name of the testing management system.  Options currently limited to Linkit and ScantronAS.}

\item{SMS}{Name of the student management system.  Options currently limited to PowerSchool.}

\item{UploadFilenames}{character of upload filenames.  The first is for percentage scores and the second is for total points.}

\item{template}{alternate template file}

\item{messageLevel}{integer - level of messages to print}
}
\value{
an object of class REPORT
}
\description{
Generate the excel score report and narrative
}
